the start menu files have been removed and refactored into a single file which was named TaskbarwStart until i renamed it to Taskbar. the App file has also been refactored
